+++ admin/modals/user/user_edit.php
@@ -58,7 +58,7 @@ ob_start();
 
                 <center class="mb-3">
                     <?php if (!empty($user_avatar)) { ?>
-                        <img class="img-fluid" src="<?php echo "../uploads/users/$user_id/$user_avatar"; ?>">
+                        <img class="img-fluid" src="<?php echo "/photo.php?type=user&id=$user_id"; ?>">
                     <?php } else { ?>
                         <span class="fa-stack fa-4x">
                             <i class="fa fa-circle fa-stack-2x text-secondary"></i>