+++ agent/modals/contact/contact_edit.php
@@ -283,7 +283,7 @@ ob_start();
 
                 <div class="mb-3 text-center">
                     <?php if ($contact_photo) { ?>
-                        <img class="img-fluid" alt="contact_photo" src="<?php echo "../uploads/clients/$client_id/$contact_photo"; ?>">
+                        <img class="img-fluid" alt="contact_photo" src="<?php echo "/photo.php?type=contact&id=$contact_id"; ?>">
                     <?php } else { ?>
                         <span class="fa-stack fa-4x">
                             <i class="fa fa-circle fa-stack-2x text-secondary"></i>