+++ client/includes/header.php
@@ -128,7 +128,7 @@ header("X-Frame-Options: DENY"); // Legacy
     <div class="row mb-3">
         <div class="col-md-1 text-center">
             <?php if (!empty($session_contact_photo)) { ?>
-                <img src="/uploads/clients/<?= $session_client_id ?>/<?= $session_contact_photo ?>" alt="..." height="50" width="50" class="img-circle img-responsive">
+                <img src="/photo.php?type=contact&id=<?= $session_contact_id ?>" alt="..." height="50" width="50" class="img-circle img-responsive">
 
             <?php } else { ?>
                 <span class="fa-stack fa-2x rounded-left">